api/requirements.txt
azure-functions==1.18.0
azure-functions-worker==1.0.0
Pillow==10.0.0
pandas==2.0.3
openpyxl==3.1.2
flask==2.3.2
flask-cors==4.0.0
